refactor(protocol): make OpenAI wire layer runtime-neutral
