--- wmo/runtime/gateway/openai/__init__.py
@@ -1,11 +1,0 @@
-"""OpenAI-compatible gateway protocol decoding, streaming, and replay state."""
-
-from wmo.runtime.gateway.openai.errors import OpenAIProtocolError
-from wmo.runtime.gateway.openai.requests import DecodedGatewayRequest, decode_chat, decode_responses
-
-__all__ = [
-    "DecodedGatewayRequest",
-    "OpenAIProtocolError",
-    "decode_chat",
-    "decode_responses",
-]
