+++ wmo/runtime/openai_protocol/__init__.py
@@ -1,0 +1,15 @@
+"""Shared OpenAI protocol decoding, streaming, and replay state."""
+
+from wmo.runtime.openai_protocol.errors import OpenAIProtocolError
+from wmo.runtime.openai_protocol.requests import (
+    DecodedGatewayRequest,
+    decode_chat,
+    decode_responses,
+)
+
+__all__ = [
+    "DecodedGatewayRequest",
+    "OpenAIProtocolError",
+    "decode_chat",
+    "decode_responses",
+]
